[DO NOT MERGE ANYWHERE] Do not set device_provisioned
Instead use the default value, which is false. Setting the value true
in the overlay file was done years ago when there was no SetupWizard in
Android TV (at least, that is what the git logs suggest).

Test: manual. This change has a smaller impact than it otherwise would,
because SetupWraith's MainActivity re-sets provisioned back to false in
its onCreate(); so before this patch there was only a 5-second or so
window when the provisioned flag had the wrong value.

b/27338096

This is for TV GMS only and will be Cherry Picked to nyc-gms-release

Change-Id: Ied83103fa1f5313262d0f8265710c10be299b22c