Disable GPS for TV SDK
TV SDK is not designed to have GPS support,
which mean that there shoudn't be any support of GNSS-receiver.
To make GNSS unavailable (which will remove GPS_PROVIDER from LocationManager)
HAL definition should be disabled and GNSS components removed.

Original change was introduced and merged here: ag/10098772
but wasn't cherry-picked directly due to the refactoring
of the manifest usage approach in master (no needs in overriding of the whole manifest).

Bug: 147219553
Test: atest SystemFeaturesTest#testLocationFeatures,CtsLocationTestCases

Change-Id: I2cc34899b036fe8199e7f8b4d81daccfe3fedb14