Lock ATV UI mode to prevent Car Mode tests from running + failing
ATV will refuse to go into any other UI mode than TV UI mode. In
a future version of Android we can just rely on this flag instead-
for now we add the flag to make tests pass, but the actual
behaviour doesn't change because it was already correct.

Change-Id: I7aa546ed6e93e5404668a2883dfeeb936bc0ab1e
Test: atest android.app.cts.UiModeManagerTest
Fix: 158358991