Fix Multiple Client Support in Audio Proxy HAL
When AudioPolicyManager first opens this HAL, it iterates through the
devices and quickly opens and closes the first device (as specified in
the audio configuration .xml). However, it is possible that the first
device has not registered with the audio proxy HAL yet. In this case, we
will return a dummy device, which is going to create a dummy output
stream. Since this HAL only supports direct outputs, the dummy output
will be immediately closed until it is reopened on use -- and by that
time the actual device must have registered itself.

Bug: 159007544
Test: Tested using test clients on ATV device (ag/11818224)
Change-Id: I17349bc0f433d06798e21f91bb2cd4aa8749df2c
(cherry picked from commit dcbdea6bb79900fe604487a016766a4bbd025b30)