Enable adb connection on user builds.
This is an emulator-only change.  In order to use
tools (e.g. cts, android studio) as expected, we need ADB
enabled by default.  Without this change, adb only works with
userdebug builds or if you go into the UI of the emulator and
enable "USB Debugging" manually.

This change makes the TV emulator work like the phone
emulator. The equivilent phone code is located at
http://cs/android/vendor/google/sdk/GoogleSdkSetup/src/com/google/android/sdksetup/DefaultActivity.java?l=59&rcl=1fe3f9a12e15fed7fa64824fb25cec55968b6b6c

Bug: b/170487716
Test: Launched the TV emulator, asserted that adb can connect.
Change-Id: I75ac073b896719a95631729098ea469e410ae740