[AudioProxy] Remove libhardware_headers dependency
The libhardware_headers dependency is introduced by
android.hardware.audio.common.util. We use the macros defined in the
header files. This patch copies the macro into AudioProxyVersionMacro.h
so that we can remove that dependency.

This allows us to move audio proxy back to product parition, where
libhardware_headers dependency is not allowed.

Bug: b/179300588
Test: Build and test on device.
Change-Id: I144a028d3b1f902f24bf9ac5ef0c9bc01b00501a