[AudioProxy] Start the AIDL service
This patch implements the logic to start the AIDL AudioProxy service and
the HIDL audio HAL service. Newly added classes:
1. AudioProxyImpl -- AIDL AudioProxy impl.
2. BusStreamProvider -- Wrapper of AIDL IStreamProvider. It's a bridge
between the AIDL IOutputStream and the HIDL IStreamOut.
3. StreamOutImpl -- audio HIDL IStreamOut impl.
4. DummyBusOutputStream -- A output stream that blocks audio data write.
This is used when the AudioProxy client doesn't provide valid
IOutputStream.

The use of AIDL IOutputStream will be added in a later patch.

Bug: 186053337
Test: Build and test on local device.
Change-Id: I2e3762a5f5cffca1723d7beb47e5e3327a696e7a