[AudioProxy] Allow client to provide customized buffer size/latency
When the server creates output stream with 0 buffer size and 0 latency,
the client should provide their buffer size and latency. This will be
used in most of the cases where the service creates a new IOutputStream
for audio HAL IStreamOut.

When the server creates output stream with valid buffer size / latency,
the client should try to follow the specified value. This is usually
used for streams running on mixer thread, where the audio HAL IStreamOut
will be opened before IOutputStream creation. The default value comes
from command line.

Bug: 232984408
Test: Build, on device test
Change-Id: I391d3b9e8bc66eba00cc484fcb60ac302bdc7ae6