Disables ethernet on ATV/GTV emulator.
ATV can't enable WiFi when ethernet is enabled, which is causing mass
test failures. To avoid this, we'll turn off ethernet for now to stop
obscuring test failures.

We'll look into turning this back on later if it's required, or editing
the CTS tests so that they don't require ethernet to be turned off.

Test: CTS tests.
Bug: 268595169
Change-Id: If9eb47ca3ec5f05c7da89be3ba99193788aab06c