Freeze vendor API
In Android 16, the "Feature Launch Flag" mechanism was introduced. As a result, Vendor APIs that have not been frozen are treated as untested code and are not permitted to be compiled or executed.
Since the mDNS offload API has been stable since Android 14, it is appropriate to promote it to a stable AIDL.

Bug: 438335571

Change-Id: I95c67866bb30d59ca5c413924ffb5cd3b5954392