Move Enable HDMI-CEC description
HDMI CEC preference description is too long, especially when text size is large, so the "Enter standby..." item pushed off the screen in such a case, so it's not obvious to the user that another option is under the text. The solution is that we need to move the text for "Enable HDMI CEC" to the right side.

Current Behavior:
https://hsv.googleplex.com/5101916383084544

Solution:
https://hsv.googleplex.com/5899310213365760

Flag: NONE not needed
Bug: 355462390
Change-Id: Ia04158bbf1a8216542625fe3139505917946a537