Fix MdnsOffloadManagerService crash without Low Power Standby
MdnsOffloadManagerService was crashing when
config_lowPowerStandbySupported is set to false due to a
NullPointerException. Add a check at service startup to see if the
feature is supported and stops the service if it is not.

The powerManager.isLowPowerStandbySupported() is related to
config_lowPowerStandbySupported, which is determined at build time. And
due to the early stop of the service, there is no need to stop the
vendor offload service when Low Power Standby is not enabled.

Test: manual and CtsNativeNetDnsTestCases pass
Bug: 448534011
Change-Id: I0ea88f7a92593909fded5ff03c089b00dca69446