Fix "null disceonncted" toast on forget BT earbuds
When a user tries to forget connected earbuds, the Bluetooth process
triggers the unbond and disconnect flows simultaneously. If the unbond
flow completes first, the earbuds information is cleared and the device
name can not be retrieved when the disconnect flow finishes. This
results in a confusing "null disconnect" toast message for the user.

Add a shared preference object to record the earbuds name when the
unbond flow begins. If we find that the earbuds' record has been
cleared, we can use the name stored in the share preference instead.

Bug: 453895260
Test: manual - Forget a connected A2DP device -> Disconnect toast
showed with correct device name
Change-Id: Ia596681fc8326ad45eae930e80aed426acbf695f